+++ HISTORY.md
@@ -4,6 +4,8 @@ History
 0.1.2 (TBD)
 -----------
 
+- Add --check-template option.
+- Add --log-silently option.
 - Switch from using optparse to argparse.
 
 0.1.1 (2012-07-01)